+++ src/components/modules/banners/contributionsTemplate/ContributionsTemplateTicker.tsx
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { css, SerializedStyles } from '@emotion/core';
 import { palette } from '@guardian/src-foundations';
 import { headline } from '@guardian/src-foundations/typography';
-import { TickerSettings } from '../../../../lib/variants';
 import useTicker from '../../../../hooks/useTicker';
 import { useHasBeenSeen, HasBeenSeen } from '../../../../hooks/useHasBeenSeen';
+import { TickerSettings } from '../../../../types/shared';
 
 // This ticker component provides an animated progress bar and counter for the
 // epic. It mirrors the behaviour of the "unlimited" ticker type from frontend.
