+++ src/components/modules/banners/contributionsTemplate/ExampleContributionsTemplate.tsx
@@ -13,7 +13,7 @@ import ContributionsTemplateSupportingText from './ContributionsTemplateSupporti
 import ContributionsTemplateTicker from './ContributionsTemplateTicker';
 import ContributionsTemplateCta from './ContributionsTemplateCta';
 import { BannerProps } from '../../../../types/BannerTypes';
-import { TickerSettings } from '../../../../lib/variants';
+import { TickerSettings } from '../../../../types/shared';
 
 const closeButtonStyles = css`
     color: ${neutral[7]};
