+++ src/components/modules/epics/ContributionsEpicButtons.tsx
@@ -2,13 +2,13 @@ import React, { useEffect } from 'react';
 import { css } from '@emotion/core';
 import { space } from '@guardian/src-foundations';
 import { Button } from './Button';
-import { EpicTracking } from './ContributionsEpicTypes';
-import { Cta, SecondaryCtaType, Variant } from '../../../lib/variants';
+import { EpicTracking, EpicVariant, SecondaryCtaType } from '../../../types/EpicTypes';
 import { addRegionIdAndTrackingParamsToSupportUrl } from '../../../lib/tracking';
 import { getCookie } from '../../../lib/cookies';
 import { OphanComponentEvent } from '../../../types/OphanTypes';
 import { getReminderViewEvent, OPHAN_COMPONENT_EVENT_REMINDER_OPEN } from './utils/ophan';
 import { useHasBeenSeen } from '../../../hooks/useHasBeenSeen';
+import { Cta } from '../../../types/shared';
 
 const buttonWrapperStyles = css`
     margin: ${space[6]}px ${space[2]}px ${space[1]}px 0;
@@ -83,7 +83,7 @@ const SecondaryCtaButton = ({
 };
 
 interface ContributionsEpicButtonsProps {
-    variant: Variant;
+    variant: EpicVariant;
     tracking: EpicTracking;
     countryCode?: string;
     onOpenReminderClick: () => void;
