+++ src/factories/tracking.ts
@@ -3,7 +3,7 @@ import {
     EpicTestTracking,
     EpicPageTracking,
     EpicTracking,
-} from '../components/modules/epics/ContributionsEpicTypes';
+} from '../types/EpicTypes';
 
 export const pageTracking = Factory.define<EpicPageTracking>(() => ({
     ophanPageId: 'k5nxn0mxg7ytwpkxuwms',
