+++ src/tests/amp/ampEpicModels.ts
@@ -1,6 +1,6 @@
-import { Cta, TickerSettings } from '../../lib/variants';
 import { CountryGroupId } from '../../lib/geolocation';
 import { AMPTicker } from './ampTicker';
+import { Cta, TickerSettings } from '../../types/shared';
 
 /**
  * Models for the data returned to AMP
