+++ src/types/BannerTypes.ts
@@ -1,9 +1,13 @@
-import { TickerSettings } from '../lib/variants';
 import { OphanProduct, OphanComponentType, OphanComponentEvent } from './OphanTypes';
 import { CountryGroupId } from '../lib/geolocation';
-import { ArticlesViewedSettings, WeeklyArticleHistory } from './shared';
-
-// TODO - it may be worth sharing some types with Epic tests
+import {
+    ArticlesViewedSettings, Audience,
+    Cta,
+    Test,
+    TickerSettings,
+    Variant,
+    WeeklyArticleHistory,
+} from './shared';
 
 export type BannerTargeting = {
     alreadyVisitedCount: number;
@@ -41,11 +45,6 @@ export type BannerDataRequestPayload = {
     targeting: BannerTargeting;
 };
 
-export interface Cta {
-    text: string;
-    baseUrl: string;
-}
-
 export interface BannerContent {
     heading?: string;
     messageText: string;
@@ -62,7 +61,7 @@ export enum BannerTemplate {
     G200Banner = 'G200Banner',
 }
 
-export interface BannerVariant {
+export interface BannerVariant extends Variant {
     name: string;
     tickerSettings?: TickerSettings;
     modulePathBuilder: (version?: string) => string;
@@ -75,18 +74,13 @@ export interface BannerVariant {
 
 export type BannerChannel = 'contributions' | 'subscriptions';
 export type CanRun = (targeting: BannerTargeting, pageTracking: BannerPageTracking) => boolean;
-export type BannerAudience =
-    | 'AllExistingSupporters'
-    | 'AllNonSupporters'
-    | 'Everyone'
-    | 'PostAskPauseSingleContributors';
 
 export type BannerTestGenerator = () => Promise<BannerTest[]>;
 
-export interface BannerTest {
+export interface BannerTest extends Test<BannerVariant> {
     name: string;
     bannerChannel: BannerChannel;
-    testAudience: BannerAudience;
+    testAudience: Audience;
     canRun: CanRun;
     minPageViews: number;
     variants: BannerVariant[];
@@ -140,7 +134,7 @@ export interface RawTestParams {
     nickname: string;
     isOn: boolean;
     minArticlesBeforeShowingBanner: number;
-    userCohort: BannerAudience;
+    userCohort: Audience;
     locations: CountryGroupId[];
     variants: RawVariantParams[];
     articlesViewedSettings?: ArticlesViewedSettings;
