+++ src/types/shared.ts
@@ -1,3 +1,49 @@
+export interface Variant {
+    name: string;
+}
+export interface Test<V extends Variant> {
+    name: string;
+    variants: V[];
+}
+
+export interface Cta {
+    text: string;
+    baseUrl: string;
+}
+
+export type Audience =
+    | 'AllExistingSupporters'
+    | 'AllNonSupporters'
+    | 'Everyone'
+    | 'PostAskPauseSingleContributors';
+
+export enum TickerEndType {
+    unlimited = 'unlimited',
+    hardstop = 'hardstop', // currently unsupported
+}
+export enum TickerCountType {
+    money = 'money',
+    people = 'people',
+}
+interface TickerCopy {
+    countLabel: string;
+    goalReachedPrimary: string;
+    goalReachedSecondary: string;
+}
+
+export interface TickerData {
+    total: number;
+    goal: number;
+}
+
+export interface TickerSettings {
+    endType: TickerEndType;
+    countType: TickerCountType;
+    currencySymbol: string;
+    copy: TickerCopy;
+    tickerData?: TickerData;
+}
+
 export type WeeklyArticleLog = {
     week: number;
     count: number;
